Remove useMenuPositioner hook
